add first content and change head in component
