--- .htaccess
@@ -1,4 +1,0 @@
-AuthType Basic
-AuthName "Restricted Access"
-AuthUserFile /.htpasswd
-Require valid-user
