--- .htpasswd
@@ -1,1 +1,0 @@
-stellan: process.env.PASSWORD;
