+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "klubklub",
+  "name": "dakiekste",
   "version": "0.0.0-unreleased",
   "private": true,
   "scripts": {
